Fix Code Style
