+++ example.txt
@@ -1,0 +1,1 @@
+[T][ ] like
