--- src/main/java/Niko/Command/ExitCommand.java
@@ -1,33 +1,0 @@
-package Niko.Command;
-
-import Niko.Main.Storage;
-import Niko.Main.Ui;
-import Niko.Task.TaskList;
-
-/**
- * Represents a command to exit the application.
- */
-public class ExitCommand extends Command {
-
-    /**
-     * Executes the exit command, showing a goodbye message to the user.
-     *
-     * @param tasks   The task list (not used in this command).
-     * @param ui      The UI to display the goodbye message.
-     * @param storage The storage (not used in this command).
-     */
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.showGoodbyeMessage();
-    }
-
-    /**
-     * Indicates that this command will exit the application.
-     *
-     * @return {@code true}, as this command exits the application.
-     */
-    @Override
-    public boolean isExit() {
-        return true;
-    }
-}
