--- src/main/java/Niko/Command/UnmarkCommand.java
@@ -1,49 +1,0 @@
-package Niko.Command;
-
-import Niko.Common.NikoException;
-import Niko.Main.Storage;
-import Niko.Main.Ui;
-import Niko.Task.TaskList;
-
-/**
- * Represents a command to unmark a task as done in the task list.
- */
-public class UnmarkCommand extends Command {
-    private int INDEX;
-
-    /**
-     * Constructs an UnmarkCommand with the specified task index.
-     *
-     * @param index The index of the task to be unmarked as done (1-based).
-     */
-    public UnmarkCommand(int index) {
-        this.INDEX = index;
-    }
-
-    /**
-     * Executes the unmark command, unmarking the task at the specified index as not done,
-     * updating the UI, and saving the updated task list to storage.
-     *
-     * @param tasks   The task list to operate on.
-     * @param ui      The UI to interact with the user.
-     * @param storage The storage to save the updated task list.
-     * @throws NikoException If an error occurs during execution.
-     */
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) throws NikoException {
-        tasks.unmarkTaskAsDone(INDEX - 1);
-        ui.showUnmarkTaskMessage(tasks.getTask(INDEX - 1));
-        String readyToWrite = tasks.getTasks().toString();
-        storage.write(readyToWrite.substring(1, readyToWrite.length() - 1));
-    }
-
-    /**
-     * Indicates that this command does not exit the application.
-     *
-     * @return {@code false}, as this command does not exit the application.
-     */
-    @Override
-    public boolean isExit() {
-        return false;
-    }
-}
