--- src/main/java/Niko/Common/NikoException.java
@@ -1,15 +1,0 @@
-package Niko.Common;
-
-/**
- * Represents an exception specific to the Niko application.
- */
-public class NikoException extends Exception {
-    /**
-     * Constructs a NikoException with the specified error message.
-     *
-     * @param message The error message describing the exception.
-     */
-    public NikoException(String message) {
-        super(message);
-    }
-}
