--- src/main/java/Niko/Main/Main.java
@@ -1,16 +1,0 @@
-package Niko.Main;
-
-/**
- * The Main class to start the Niko chatbot.
- */
-public class Main {
-    /**
-     * The main method to start the Niko chatbot.
-     *
-     * @param args Command line arguments.
-     */
-    public static void main(String[] args) {
-        Niko chatbot = new Niko("example.txt");
-        chatbot.run();
-    }
-}
