--- src/main/java/Niko/Task/TaskStatus.java
@@ -1,5 +1,0 @@
-package Niko.Task;
-
-public enum TaskStatus {
-    DONE, NOT_DONE
-}
