--- src/main/java/Niko/Task/TaskType.java
@@ -1,5 +1,0 @@
-package Niko.Task;
-
-public enum TaskType {
-    TODO, DEADLINE, EVENT
-}
