--- src/main/java/Niko/Task/Todo.java
@@ -1,21 +1,0 @@
-package Niko.Task;
-
-/**
- * Represents a task of type TODO.
- */
-public class Todo extends Task {
-
-    /**
-     * Constructs a Todo task with the specified description.
-     *
-     * @param description The description of the task.
-     */
-    public Todo(String description) {
-        super(description, TaskType.TODO);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
