+++ src/main/java/niko/task/TaskStatus.java
@@ -1,0 +1,10 @@
+package niko.task;
+
+/**
+ * Represents the status of a task.
+ * A task can either be marked as DONE or NOT_DONE.
+ */
+
+public enum TaskStatus {
+    DONE, NOT_DONE
+}
