+++ src/main/java/niko/task/TaskType.java
@@ -1,0 +1,11 @@
+package niko.task;
+
+
+/**
+ * Represents the type of a task.
+ * The task can be one of the following types: TODO, DEADLINE, or EVENT.
+ */
+
+public enum TaskType {
+    TODO, DEADLINE, EVENT
+}
